Crate minifb

Reexports

pub use key::Key;

Modules

key

os

Structs

Menu

Window

WindowOptions

Enums

Error
Errors that can be return from various operatiors

KeyRepeat
Used for is_key_pressed and get_keys_pressed() to indicated if repeat of presses is wanted

MouseButton
The various mouse buttons that are availible

MouseMode
Key is used by the get key functions to check if some keys on the keyboard has been pressed

Scale
Scale will scale the frame buffer and the window that is being sent in when calling the update function. This is useful if you for example want to display a 320 x 256 window on a screen with much higher resolution which would result in that the window is very small.

Constants

MENU_KEY_ALT
Alt key

MENU_KEY_COMMAND
Command key on Mac OS

MENU_KEY_CTRL
Control key

MENU_KEY_SHIFT
Shift key

MENU_KEY_WIN
Windows key on Windows

Type Definitions

Result