Enum minifb::Scale

pub enum Scale {
    FitScreen,
    X1,
    X2,
    X4,
    X8,
    X16,
    X32,
}

Scale will scale the frame buffer and the window that is being sent in when calling the update function. This is useful if you for example want to display a 320 x 256 window on a screen with much higher resolution which would result in that the window is very small.

Variants

FitScreen

This mode checks your current screen resolution and will caluclate the largest window size that can be used within that limit and resize it. Useful if you have a small buffer to display on a high resolution screen.

X1

1X scale (which means leave the corrdinates sent into Window::new untouched)

X2

2X window scale (Example: 320 x 200 -> 640 x 400)

X4

4X window scale (Example: 320 x 200 -> 1280 x 800)

X8

8X window scale (Example: 320 x 200 -> 2560 x 1600)

X16

16X window scale (Example: 320 x 200 -> 5120 x 3200)

X32

32 window scale (Example: 320 x 200 -> 10240 x 6400)