Struct minifb::UnixMenuItem

pub struct UnixMenuItem {
    pub sub_menu: Option<Box<UnixMenu>>,
    pub handle: MenuItemHandle,
    pub id: usize,
    pub label: String,
    pub enabled: bool,
    pub key: Key,
    pub modifier: usize,
}

Used for on Unix (Linux, FreeBSD, etc) as menus aren't supported in a native where there. This structure holds info for each item in a #UnixMenu

Fields
sub_menu: Option<Box<UnixMenu>>

Set to a menu if there is a Item is a sub_menu otherwise None

handle: MenuItemHandle

Handle of the MenuItem

id: usize

Id of the item (set by the user from the outside and should be reported back when pressed)

label: String

Name of the item

enabled: bool

Set to true if enabled otherwise false

key: Key

Shortcut key

modifier: usize

Modifier for the key (Shift, Ctrl, etc)