Struct minifb::WindowOptions

pub struct WindowOptions {
    pub borderless: bool,
    pub title: bool,
    pub resize: bool,
    pub scale: Scale,
}

WindowOptions is creation settings for the window. By default the settings are defined for displayng a 32-bit buffer (no scaling of window is possible)

Fields

borderless

If the window should be borderless (default: false)

title

If the window should have a title (default: true)

resize

If it should be possible to resize the window (default: false)

scale

Scale of the window that used in conjunction with update_with_buffer (default: X1)